Update AI alert scoring and template presentation
This commit introduces two key changes to the ai_alerts and the alerts.jinja2 template.

Firstly, in the main.py file, a new line of code has been inserted to round and transform the 'score' field of ai_alerts into a percentage (multiplied by 100). The purpose of this change is to present a more understandable score to users when showing the confidence level of ai_alerts.

Secondly, major changes were made to the alerts.jinja2 template to enhance the presentation of alerts. The template now leads with BlazeWatch Wildfire Forecast Predictions, which display the AI predictions before showing the Latest Wildfire Reports. These changes aim to prioritize presentation of AI-predicted fires over manually reported ones and make the alerts more comprehensible to users.